+++ packages/plugins/CHANGELOG.md
@@ -1,5 +1,7 @@
 # seroval-plugins
 
+## 1.6.2
+
 ## 1.6.1
 
 ### Patch Changes